Refactor: remove CaseFileViewRead component
