--- projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-file-view/case-file-view-field-read.component.spec.ts
@@ -1,100 +1,0 @@
-import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { ActivatedRoute } from '@angular/router';
-import { RouterTestingModule } from '@angular/router/testing';
-import { of } from 'rxjs';
-import { CaseField } from '../../../domain';
-import { CaseFileViewService, DocumentManagementService, LoadingService, SessionStorageService } from '../../../services';
-import { mockDocumentManagementService } from '../../../services/document-management/document-management.service.mock';
-import { CaseFileViewFieldReadComponent } from './case-file-view-field-read.component';
-import createSpyObj = jasmine.createSpyObj;
-
-describe('CaseFileViewFieldReadComponent', () => {
-  let component: CaseFileViewFieldReadComponent;
-  let fixture: ComponentFixture<CaseFileViewFieldReadComponent>;
-  let mockCaseFileViewService: jasmine.SpyObj<CaseFileViewService>;
-  let mockLoadingService: jasmine.SpyObj<LoadingService>;
-  let mockSessionStorageService: jasmine.SpyObj<SessionStorageService>;
-
-  const mockSnapshot = {
-    paramMap: createSpyObj('paramMap', ['get']),
-  };
-  const mockRoute = {
-    params: of({ cid: '1234123412341234' }),
-    snapshot: mockSnapshot
-  };
-  const mockCaseFieldTrue = {
-    acls: [
-      {
-        create: true,
-        read: true,
-        update: true,
-        delete: false,
-        role: 'caseworker-privatelaw-judge'
-      }
-    ]
-  };
-  const mockCaseFieldFalse = {
-    acls: [
-      {
-        create: true,
-        read: true,
-        update: false,
-        delete: false,
-        role: 'caseworker-privatelaw-judge'
-      },
-    ]
-  };
-  const mockUser: string = JSON.stringify({ roles: ['caseworker-privatelaw-judge'] });
-
-  beforeEach(() => {
-    mockCaseFileViewService = createSpyObj<CaseFileViewService>('CaseFileViewService', ['getCategoriesAndDocuments']);
-    mockCaseFileViewService.getCategoriesAndDocuments.and.returnValue(of(null));
-
-    mockLoadingService = createSpyObj<LoadingService>('LoadingService', ['register', 'unregister']);
-    mockLoadingService.register.and.returnValue('loadingToken');
-    mockLoadingService.unregister.and.returnValue(null);
-
-    mockSessionStorageService = createSpyObj<SessionStorageService>('SessionStorageService', ['getItem']);
-    mockSessionStorageService.getItem.and.returnValue(mockUser);
-
-    TestBed.configureTestingModule({
-      imports: [
-        RouterTestingModule
-      ],
-      declarations: [
-        CaseFileViewFieldReadComponent
-      ],
-      schemas: [
-        CUSTOM_ELEMENTS_SCHEMA
-      ],
-      providers: [
-        { provide: ActivatedRoute, useValue: mockRoute },
-        { provide: CaseFileViewService, useValue: mockCaseFileViewService },
-        { provide: DocumentManagementService, useValue: mockDocumentManagementService },
-        { provide: LoadingService, useValue: mockLoadingService },
-        { provide: SessionStorageService, useValue: mockSessionStorageService },
-
-      ]
-    });
-    fixture = TestBed.createComponent(CaseFileViewFieldReadComponent);
-    component = fixture.componentInstance;
-    component.caseField = new CaseField();
-  });
-
-  it('should set allowMoving to true based on acl update being true', () => {
-    component.caseField.acls = mockCaseFieldTrue.acls;
-
-    fixture.detectChanges();
-
-    expect(component.allowMoving).toBeTruthy();
-  });
-
-  it('should set allowMoving to false based on acl update being false', () => {
-    component.caseField.acls = mockCaseFieldFalse.acls;
-
-    fixture.detectChanges();
-
-    expect(component.allowMoving).toBeFalsy();
-  });
-});
