+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-file-view/case-file-view-field.component.spec.ts
@@ -9,6 +9,7 @@ import { CaseFileViewService, DocumentManagementService, LoadingService } from '
 import { mockDocumentManagementService } from '../../../services/document-management/document-management.service.mock';
 import createSpyObj = jasmine.createSpyObj;
 import { CaseFileViewFieldComponent } from './case-file-view-field.component';
+import { CaseField } from '../../../domain';
 
 describe('CaseFileViewFieldComponent', () => {
   let component: CaseFileViewFieldComponent;
@@ -25,6 +26,30 @@ describe('CaseFileViewFieldComponent', () => {
     snapshot: mockSnapshot
   };
 
+  const mockCaseFieldTrue = {
+    acls: [
+      {
+        create: true,
+        read: true,
+        update: true,
+        delete: false,
+        role: 'caseworker-privatelaw-judge'
+      }
+    ]
+  };
+  const mockCaseFieldFalse = {
+    acls: [
+      {
+        create: true,
+        read: true,
+        update: false,
+        delete: false,
+        role: 'caseworker-privatelaw-judge'
+      },
+    ]
+  };
+  const mockUser: string = JSON.stringify({ roles: ['caseworker-privatelaw-judge'] });
+
   beforeEach(async(() => {
     mockCaseFileViewService = createSpyObj<CaseFileViewService>('CaseFileViewService', ['getCategoriesAndDocuments']);
     mockCaseFileViewService.getCategoriesAndDocuments.and.returnValue(of(null));
@@ -34,6 +59,7 @@ describe('CaseFileViewFieldComponent', () => {
     mockLoadingService.unregister.and.returnValue(null);
 
     mockSessionStorageService = createSpyObj<SessionStorageService>('SessionStorageService', ['getItem']);
+    mockSessionStorageService.getItem.and.returnValue(mockUser);
 
     TestBed.configureTestingModule({
       imports: [
@@ -59,10 +85,12 @@ describe('CaseFileViewFieldComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(CaseFileViewFieldComponent);
     component = fixture.componentInstance;
-    fixture.detectChanges();
+    component.caseField = new CaseField();
+    component.caseField.acls = mockCaseFieldTrue.acls;
   });
 
   it('should create component', () => {
+    fixture.detectChanges();
     expect(component).toBeTruthy();
     expect(mockCaseFileViewService.getCategoriesAndDocuments).toHaveBeenCalled();
     expect(component.getCategoriesAndDocumentsError).toBe(false);
@@ -105,5 +133,17 @@ describe('CaseFileViewFieldComponent', () => {
 
     expect(component.currentDocument.document_filename).toEqual(dummyNodeTreeDocument.document_filename);
     expect(component.currentDocument.document_binary_url).toEqual(dummyNodeTreeDocument.document_binary_url);
+  });
+
+  it('should set allowMoving to true based on acl update being true', () => {
+    component.caseField.acls = mockCaseFieldTrue.acls;
+    fixture.detectChanges();
+    expect(component.allowMoving).toBeTruthy();
+  });
+
+  it('should set allowMoving to false based on acl update being false', () => {
+    component.caseField.acls = mockCaseFieldFalse.acls;
+    fixture.detectChanges();
+    expect(component.allowMoving).toBeFalsy();
   });
 });
