+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-file-view/index.ts
@@ -1,3 +1,2 @@
-export * from './case-file-view-field-read.component';
 export * from './case-file-view-field.component';
 export * from './components';
