+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/palette.module.ts
@@ -32,7 +32,6 @@ import { WindowService } from '../../services/window';
 import { WriteAddressFieldComponent } from './address/write-address-field.component';
 import { FieldReadComponent, FieldReadLabelComponent, FieldWriteComponent } from './base-field';
 import { CaseFileViewOverlayMenuComponent } from './case-file-view';
-import { CaseFileViewFieldReadComponent } from './case-file-view/case-file-view-field-read.component';
 import { CaseFileViewFieldComponent } from './case-file-view/case-file-view-field.component';
 import { CaseFileViewFolderSelectorComponent } from './case-file-view/components/case-file-view-folder-selector/case-file-view-folder-selector.component';
 import { CaseFileViewFolderDocumentActionsComponent } from './case-file-view/components/case-file-view-folder/case-file-view-folder-document-actions/case-file-view-folder-document-actions.component';
@@ -160,7 +159,6 @@ const PALETTE_COMPONENTS = [
 
   // ComponentLauncher web components
   CaseFileViewFieldComponent,
-  CaseFileViewFieldReadComponent,
   CaseFileViewFolderComponent,
   CaseFileViewFolderSortComponent,
   CaseFileViewOverlayMenuComponent,
